various bugfixes
git-svn-id: svn://svn.code.sf.net/p/szte-wsn/code/tinyos-2.x/trunk@852 a4e16f3a-606c-4714-833f-ed604d7aca08

+++ support/java/org/szte/wsn/dataprocess/parser/Taos2550Parser.java
@@ -30,7 +30,7 @@ public class Taos2550Parser extends IntegerParser{
 	{	
 		int light=byteToCount(packet[0]);
 
-		return new String[] {Double.toString(light)};	
+		return new String[] {Integer.toString(light)};	
 	}
 	
 	/**
